+++ ui/mods/instant_sandbox/settings.js
@@ -38,8 +38,6 @@
 
   }
 
-  $('.slider').attr('data-slider-tooltip', 'show')
-
   var instant_sandbox_settings = {
     instant_sandbox_play: {
       title: 'Play',
@@ -102,8 +100,11 @@
   var $group = $('<div class="sub-group"></div>').appendTo('.option-list.ui .form-group')
   $group.append('<div class="sub-group-title">Instant Sandbox</div>')
 
+  var $template = $('script#setting-template')
+  $group.append($template[0].outerHTML.replace('setting-template', 'instant-sandbox-setting-template').replace('hide', 'show'))
+
   Object.keys(instant_sandbox_settings).forEach(function(setting) {
-    $group.append('<div class="option" data-bind="template: { name: \'setting-template\', data: $root.settingsItemMap()[\'ui.' + setting + '\'] }"></div>')
+    $group.append('<div class="option" data-bind="template: { name: \'instant-sandbox-setting-template\', data: $root.settingsItemMap()[\'ui.' + setting + '\'] }"></div>')
   })
 
   $group.append(